Adjusted to latest version in tools
