+++ lib/Headers.groovy
@@ -41,4 +41,3 @@ class Headers {
         )
     }
 }
-
